refactor: simplify usage dashboard
