+++ apps/web/src/components/usage/usage-dashboard-page.tsx
@@ -1,22 +1,5 @@
-import {
-  BarElement,
-  CategoryScale,
-  Chart as ChartJS,
-  Legend,
-  LinearScale,
-  type ScriptableContext,
-  Tooltip,
-  type TooltipItem,
-} from 'chart.js';
 import { BarChart3Icon } from 'lucide-react';
-import * as React from 'react';
-import { Bar } from 'react-chartjs-2';
 
-import { keepPreviousData, useQuery, useSuspenseQuery } from '@tanstack/react-query';
-
-import { USAGE_DATE_RANGES, USAGE_SOURCES, type UsageDateRange } from '@stitch/shared/usage/types';
-
-import { Card, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import {
   Page,
   PageContent,
@@ -26,363 +9,13 @@ import {
   PageIcon,
   PageTitle,
 } from '@/components/ui/page';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import { enabledProviderModelsQueryOptions } from '@/lib/queries/providers';
-import { usageDashboardQueryOptions } from '@/lib/queries/usage';
-
-ChartJS.register(CategoryScale, LinearScale, BarElement, Tooltip, Legend);
-
-const ALL_FILTER = 'all';
-
-const RANGE_LABELS: Record<UsageDateRange, string> = {
-  '7d': 'Last 7 days',
-  '30d': 'Last 30 days',
-  '90d': 'Last 90 days',
-  '1y': 'Last 1 year',
-  all: 'All time',
-};
-
-const SOURCE_LABELS: Record<string, string> = {
-  chat: 'Chat',
-  automation: 'Automation',
-  automation_generation: 'Automation Generation',
-  title_generation: 'Title Generation',
-  memory_extraction: 'Memory',
-  transcription: 'Transcription',
-  recording_analysis: 'Recording Analysis',
-};
-
-const FALLBACK_SOURCE_COLORS = ['#f97316', '#10b981', '#3b82f6', '#f59e0b', '#ec4899', '#14b8a6'];
-
-const SOURCE_COLOR_CONFIG: Record<string, { cssVar: string; fallback: string }> = {
-  chat: { cssVar: '--chart-1', fallback: '#f97316' },
-  automation: { cssVar: '--chart-2', fallback: '#10b981' },
-  automation_generation: { cssVar: '--chart-5', fallback: '#8b5cf6' },
-  title_generation: { cssVar: '--chart-3', fallback: '#3b82f6' },
-  transcription: { cssVar: '--chart-4', fallback: '#f59e0b' },
-  memory_extraction: { cssVar: '--chart-4', fallback: '#ec4899' },
-  recording_analysis: { cssVar: '--chart-5', fallback: '#14b8a6' },
-};
-
-function getSourceLabel(source: string): string {
-  return SOURCE_LABELS[source] ?? source.replaceAll('_', ' ');
-}
-
-function formatCost(costUsd: number): string {
-  return `$${costUsd.toFixed(2)}`;
-}
-
-function formatTokens(value: number): string {
-  return new Intl.NumberFormat(undefined, { notation: 'compact', maximumFractionDigits: 0 }).format(
-    value,
-  );
-}
-
-function encodeModelFilter(providerId: string, modelId: string): string {
-  return `${providerId}::${modelId}`;
-}
-
-function decodeModelFilter(value: string): { providerId: string; modelId: string } | null {
-  const separator = value.indexOf('::');
-  if (separator <= 0) return null;
-  return {
-    providerId: value.slice(0, separator),
-    modelId: value.slice(separator + 2),
-  };
-}
-
-function resolveCssVar(varName: string, fallback: string): string {
-  if (typeof window === 'undefined') return fallback;
-  const value = getComputedStyle(document.documentElement).getPropertyValue(varName).trim();
-  return value.length > 0 ? value : fallback;
-}
-
-function hashSource(source: string): number {
-  let hash = 0;
-  for (let i = 0; i < source.length; i += 1) {
-    hash = (hash * 31 + source.charCodeAt(i)) >>> 0;
-  }
-  return hash;
-}
-
-function getSourceColor(source: string): string {
-  const configured = SOURCE_COLOR_CONFIG[source];
-  if (configured) {
-    return resolveCssVar(configured.cssVar, configured.fallback);
-  }
-
-  const fallback = FALLBACK_SOURCE_COLORS[hashSource(source) % FALLBACK_SOURCE_COLORS.length];
-  return fallback ?? '#6b7280';
-}
-
-function getNumericValue(value: unknown): number {
-  if (typeof value === 'number') return value;
-  if (typeof value === 'string') {
-    const parsed = Number(value);
-    return Number.isFinite(parsed) ? parsed : 0;
-  }
-  return 0;
-}
-
-function getStackSegmentRadius(ctx: ScriptableContext<'bar'>, radius = 5) {
-  const datasetIndex = ctx.datasetIndex;
-  const dataIndex = ctx.dataIndex;
-  const datasets = ctx.chart.data.datasets;
-
-  const hasAbove = datasets
-    .slice(datasetIndex + 1)
-    .some((dataset) => getNumericValue(dataset.data?.[dataIndex]) > 0);
-  const hasBelow = datasets
-    .slice(0, datasetIndex)
-    .some((dataset) => getNumericValue(dataset.data?.[dataIndex]) > 0);
-
-  return {
-    topLeft: hasAbove ? 0 : radius,
-    topRight: hasAbove ? 0 : radius,
-    bottomLeft: hasBelow ? 0 : radius,
-    bottomRight: hasBelow ? 0 : radius,
-  };
-}
-
-function useSourceOrder(sources: string[]): string[] {
-  return React.useMemo(() => {
-    const order = new Map<string, number>(USAGE_SOURCES.map((source, index) => [source, index]));
-    return [...sources].sort((a, b) => {
-      const aOrder = order.get(a) ?? Number.MAX_SAFE_INTEGER;
-      const bOrder = order.get(b) ?? Number.MAX_SAFE_INTEGER;
-      if (aOrder !== bOrder) return aOrder - bOrder;
-      return a.localeCompare(b);
-    });
-  }, [sources]);
-}
-
-function useChartTheme() {
-  return React.useMemo(() => {
-    const tickColor = resolveCssVar('--muted-foreground', '#71717a');
-    const gridColor = resolveCssVar('--border', '#27272a');
-    return { tickColor, gridColor };
-  }, []);
-}
-
-function EmptyChart({ message }: { message: string }) {
-  return (
-    <div className="flex h-full flex-col items-center justify-center gap-3 text-center">
-      <BarChart3Icon className="size-8 text-muted-foreground/40" />
-      <div>
-        <p className="text-sm font-medium text-foreground/60">No data</p>
-        <p className="text-xs text-muted-foreground">{message}</p>
-      </div>
-    </div>
-  );
-}
+import { UsageDashboardCostChart } from '@/components/usage/usage-dashboard-cost-chart';
+import { UsageDashboardFilters } from '@/components/usage/usage-dashboard-filters';
+import { UsageDashboardSummaryCards } from '@/components/usage/usage-dashboard-summary-cards';
+import { useUsageDashboardData } from '@/components/usage/use-usage-dashboard-data';
 
 export function UsageDashboardPage() {
-  const { data: providerModels } = useSuspenseQuery(enabledProviderModelsQueryOptions);
-
-  const [providerFilter, setProviderFilter] = React.useState<string>(ALL_FILTER);
-  const [modelFilter, setModelFilter] = React.useState<string>(ALL_FILTER);
-  const [rangeFilter, setRangeFilter] = React.useState<UsageDateRange>('30d');
-
-  const { data: usageRangeData } = useQuery({
-    ...usageDashboardQueryOptions({ range: rangeFilter }),
-    placeholderData: keepPreviousData,
-  });
-
-  const providerById = React.useMemo(
-    () => new Map(providerModels.map((provider) => [provider.providerId, provider] as const)),
-    [providerModels],
-  );
-
-  const modelNameByKey = React.useMemo(() => {
-    const map = new Map<string, string>();
-    for (const provider of providerModels) {
-      for (const model of provider.models) {
-        map.set(encodeModelFilter(provider.providerId, model.id), model.name);
-      }
-    }
-    return map;
-  }, [providerModels]);
-
-  const availableProviders = React.useMemo(() => {
-    const used = new Set(usageRangeData?.usedProviders ?? []);
-    return providerModels
-      .filter((provider) => used.has(provider.providerId))
-      .map((provider) => ({
-        providerId: provider.providerId,
-        providerName: provider.providerName,
-      }));
-  }, [providerModels, usageRangeData?.usedProviders]);
-
-  const availableModels = React.useMemo(() => {
-    const usedModels = usageRangeData?.usedModels ?? [];
-    return usedModels
-      .filter((model) => providerFilter === ALL_FILTER || model.providerId === providerFilter)
-      .map((model) => {
-        const provider = providerById.get(model.providerId);
-        const key = encodeModelFilter(model.providerId, model.modelId);
-        return {
-          providerId: model.providerId,
-          providerName: provider?.providerName ?? model.providerId,
-          modelId: model.modelId,
-          modelName: modelNameByKey.get(key) ?? model.modelId,
-        };
-      });
-  }, [modelNameByKey, providerById, providerFilter, usageRangeData?.usedModels]);
-
-  React.useEffect(() => {
-    if (providerFilter === ALL_FILTER) return;
-    const stillAvailable = availableProviders.some((p) => p.providerId === providerFilter);
-    if (!stillAvailable) {
-      setProviderFilter(ALL_FILTER);
-      setModelFilter(ALL_FILTER);
-    }
-  }, [availableProviders, providerFilter]);
-
-  const providerLabelById = React.useMemo(
-    () => new Map(availableProviders.map((p) => [p.providerId, p.providerName] as const)),
-    [availableProviders],
-  );
-
-  const modelLabelByValue = React.useMemo(
-    () =>
-      new Map(
-        availableModels.map(
-          (model) =>
-            [
-              encodeModelFilter(model.providerId, model.modelId),
-              `${model.providerName} · ${model.modelName}`,
-            ] as const,
-        ),
-      ),
-    [availableModels],
-  );
-
-  React.useEffect(() => {
-    if (modelFilter === ALL_FILTER) return;
-    const isStillAvailable = availableModels.some(
-      (m) => encodeModelFilter(m.providerId, m.modelId) === modelFilter,
-    );
-    if (!isStillAvailable) setModelFilter(ALL_FILTER);
-  }, [availableModels, modelFilter]);
-
-  const usageFilters = React.useMemo(() => {
-    const decodedModel = modelFilter === ALL_FILTER ? null : decodeModelFilter(modelFilter);
-    const providerIdFromModel =
-      providerFilter === ALL_FILTER ? decodedModel?.providerId : providerFilter;
-    return {
-      range: rangeFilter,
-      providerId: providerIdFromModel,
-      modelId: decodedModel?.modelId,
-    };
-  }, [modelFilter, providerFilter, rangeFilter]);
-
-  const { data: usageData, isFetching } = useQuery({
-    ...usageDashboardQueryOptions(usageFilters),
-    placeholderData: keepPreviousData,
-  });
-
-  const sources = useSourceOrder(usageData?.sources ?? []);
-  const sourceColors = React.useMemo(
-    () => Object.fromEntries(sources.map((source) => [source, getSourceColor(source)])),
-    [sources],
-  );
-  const { tickColor, gridColor } = useChartTheme();
-
-  const selectedProviderLabel =
-    providerFilter === ALL_FILTER
-      ? 'All providers'
-      : (providerLabelById.get(providerFilter) ?? 'Provider');
-  const selectedModelLabel =
-    modelFilter === ALL_FILTER ? 'All models' : (modelLabelByValue.get(modelFilter) ?? 'Model');
-  const selectedRangeLabel = RANGE_LABELS[rangeFilter];
-  const granularityLabel = usageData?.range.granularity ?? 'day';
-
-  const costChartData = React.useMemo(
-    () => ({
-      labels: usageData?.buckets.map((b) => b.label) ?? [],
-      datasets: sources.map((source) => ({
-        label: getSourceLabel(source),
-        data: usageData?.buckets.map((b) => b.costUsdBySource[source] ?? 0) ?? [],
-        backgroundColor: sourceColors[source] ?? '#6b7280',
-        borderRadius: (ctx: ScriptableContext<'bar'>) => getStackSegmentRadius(ctx),
-        borderSkipped: false,
-        inflateAmount: 0,
-      })),
-    }),
-    [sourceColors, sources, usageData],
-  );
-
-  const baseScales = React.useMemo(
-    () => ({
-      x: {
-        stacked: true,
-        grid: { display: false },
-        ticks: { color: tickColor },
-        border: { color: gridColor },
-      },
-      y: {
-        stacked: true,
-        beginAtZero: true,
-        grid: { color: gridColor },
-        ticks: { color: tickColor },
-        border: { display: false },
-      },
-    }),
-    [gridColor, tickColor],
-  );
-
-  const baseLegend = React.useMemo(
-    () => ({
-      position: 'bottom' as const,
-      labels: {
-        usePointStyle: true,
-        pointStyle: 'rectRounded' as const,
-        color: tickColor,
-        padding: 16,
-        font: { size: 12 },
-      },
-    }),
-    [tickColor],
-  );
-
-  const costChartOptions = React.useMemo(
-    () => ({
-      responsive: true,
-      maintainAspectRatio: false,
-      interaction: { mode: 'index' as const, intersect: false },
-      plugins: {
-        legend: baseLegend,
-        tooltip: {
-          callbacks: {
-            label: (ctx: TooltipItem<'bar'>) => {
-              const value = Number(ctx.raw ?? 0);
-              if (value === 0) return [];
-              return `${ctx.dataset.label}: ${formatCost(value)}`;
-            },
-          },
-        },
-      },
-      scales: {
-        ...baseScales,
-        y: {
-          ...baseScales.y,
-          ticks: {
-            ...baseScales.y.ticks,
-            callback: (value: string | number) => formatCost(Number(value)),
-          },
-        },
-      },
-    }),
-    [baseScales, baseLegend],
-  );
-
-  const hasData = !!usageData && usageData.buckets.length > 0;
+  const dashboard = useUsageDashboardData();
 
   return (
     <Page className="thin-scrollbar">
@@ -401,109 +34,23 @@ export function UsageDashboardPage() {
           </PageHeaderContent>
         </PageHeader>
 
-        {/* Filter toolbar */}
-        <div className="rounded-xl bg-muted/40 p-3">
-          <div className="grid grid-cols-1 gap-3 md:grid-cols-3">
-            <Select
-              value={providerFilter}
-              onValueChange={(value) => setProviderFilter(value ?? ALL_FILTER)}
-            >
-              <SelectTrigger className="w-full bg-background">
-                <SelectValue placeholder="Filter by provider">{selectedProviderLabel}</SelectValue>
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value={ALL_FILTER}>All providers</SelectItem>
-                {availableProviders.map((provider) => (
-                  <SelectItem key={provider.providerId} value={provider.providerId}>
-                    {provider.providerName}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
+        <UsageDashboardFilters
+          availableModels={dashboard.availableModels}
+          availableProviders={dashboard.availableProviders}
+          filters={dashboard.filters}
+          labels={dashboard.labels}
+          isFetching={dashboard.isFetching}
+          onModelChange={dashboard.setModelFilter}
+          onProviderChange={dashboard.setProviderFilter}
+          onRangeChange={dashboard.setRangeFilter}
+        />
 
-            <Select
-              value={modelFilter}
-              onValueChange={(value) => setModelFilter(value ?? ALL_FILTER)}
-            >
-              <SelectTrigger className="w-full bg-background">
-                <SelectValue placeholder="Filter by model">{selectedModelLabel}</SelectValue>
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value={ALL_FILTER}>All models</SelectItem>
-                {availableModels.map((model) => (
-                  <SelectItem
-                    key={`${model.providerId}:${model.modelId}`}
-                    value={encodeModelFilter(model.providerId, model.modelId)}
-                  >
-                    {model.providerName} · {model.modelName}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
+        <UsageDashboardSummaryCards
+          rangeLabel={dashboard.labels.range}
+          usageData={dashboard.usageData}
+        />
 
-            <div className="flex items-center gap-2">
-              <Select value={rangeFilter} onValueChange={(value) => setRangeFilter(value ?? '30d')}>
-                <SelectTrigger className="w-full bg-background">
-                  <SelectValue placeholder="Select date range">{selectedRangeLabel}</SelectValue>
-                </SelectTrigger>
-                <SelectContent>
-                  {USAGE_DATE_RANGES.map((range) => (
-                    <SelectItem key={range} value={range}>
-                      {RANGE_LABELS[range]}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-
-              {isFetching ? (
-                <div className="size-4 shrink-0 animate-spin rounded-full border-2 border-muted-foreground/30 border-t-muted-foreground" />
-              ) : null}
-            </div>
-          </div>
-        </div>
-
-        {/* KPI cards */}
-        <div className="grid grid-cols-1 gap-3 sm:grid-cols-2">
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardDescription>Total Cost</CardDescription>
-              <CardTitle className="text-3xl font-bold tabular-nums">
-                {formatCost(usageData?.totals.costUsd ?? 0)}
-              </CardTitle>
-              <p className="text-xs text-muted-foreground capitalize">
-                {selectedRangeLabel} · {granularityLabel} buckets
-              </p>
-            </CardHeader>
-          </Card>
-
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardDescription>Total Tokens</CardDescription>
-              <CardTitle className="text-3xl font-bold tabular-nums">
-                {formatTokens(usageData?.totals.tokenMetrics.totalTokens ?? 0)}
-              </CardTitle>
-              <p className="text-xs text-muted-foreground">
-                {formatTokens(usageData?.totals.tokenMetrics.inputTokens ?? 0)} in ·{' '}
-                {formatTokens(usageData?.totals.tokenMetrics.outputTokens ?? 0)} out
-              </p>
-            </CardHeader>
-          </Card>
-        </div>
-
-        {/* Cost chart */}
-        <div className="rounded-xl bg-muted/20 p-4">
-          <div className="mb-4">
-            <p className="text-sm font-medium">Cost over time</p>
-            <p className="text-xs text-muted-foreground">Stacked by source</p>
-          </div>
-          <div className="h-64 md:h-96">
-            {hasData ? (
-              <Bar data={costChartData} options={costChartOptions} />
-            ) : (
-              <EmptyChart message="No usage data for the selected filters." />
-            )}
-          </div>
-        </div>
+        <UsageDashboardCostChart usageData={dashboard.usageData} />
       </PageContent>
     </Page>
   );
